+++ Source/EventFlow/Extensions/TypeExtensions.cs
@@ -122,7 +122,12 @@ namespace EventFlow.Extensions
                 .GetMethods(BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance)
                 .Where(mi =>
                     {
-                        if (mi.Name != "Apply") return false;
+                        if (!string.Equals(mi.Name, "Apply", StringComparison.Ordinal) &&
+                            !mi.Name.EndsWith(".Apply", StringComparison.Ordinal))
+                        {
+                            return false;
+                        }
+
                         var parameters = mi.GetParameters();
                         return
                             parameters.Length == 1 &&
@@ -130,7 +135,7 @@ namespace EventFlow.Extensions
                     })
                 .ToDictionary(
                     mi => mi.GetParameters()[0].ParameterType,
-                    mi => ReflectionHelper.CompileMethodInvocation<Action<T, IAggregateEvent>>(type, "Apply", mi.GetParameters()[0].ParameterType));
+                    mi => ReflectionHelper.CompileMethodInvocation<Action<T, IAggregateEvent>>(type, mi.Name, mi.GetParameters()[0].ParameterType));
         }
     }
 }